Remove DeploySteps.txt
I don't think there's really a need for this anymore. DeploySteps
was really only here for one reason only: to make a checklist so I
don't forget to deploy to a service. That's no irrelevant because I'm
no longer deploying to so many. Now, I'm only deploying to two: Nekoweb
and GitHub Pages; of which, only 1 is a manual deploy.