WebP in blog post, button, frontpage marquee, styled inputs
I've updated the image in html-tag-ranking to use webp instead of PNG.
The reason for such a change is that WebP is much smaller than PNG in
most situations.

I've added a 88x31 button to my sidebar for people to link to my
website.

A marquee has been added to the frontpage to show sites that I'm linking
to using their provided button(or a custom one if they don't have one :P).

Inputs are now styled. This was for the button, as I'm using a readonly
textarea element, so I just decided to also change the way the other
inputs display (checkboxes, radios, etc.).