Made all pages initially load black
My site is entirely in dark theme, and since browsers often take a
couple hundred miliseconds to load CSS, it becomes painfully obvious
that the site is not fully loaded. I do not like this look. It
flashbangs the user, as some might say.

I do not want visiters to my site to be blinded, so I'm making the page
load dark using a style tag at the top of the page. It gets completely
overwritten when the actual stylesheet loads in, so it doesn't impact
the actual site's usage.