Marquee Revamp
The frontpage marquee was jank before, thanks to a bit of help from
Tropical <tropical.nekoweb.org>, it is no longer jank. That marquee was
causing me physical pain 🙃

Also, the XCF (used in GIMP) file for my Nekoweb cursor is here, it's
not used on the site, just for Nekoweb.org.

I will put a small blog post on my website about the marquee on the
frontpage so that maybe some other people can find that and be able to
it.